Rename can_scraper_providers.py to can_scraper_local_dashboard_providers.py
