+++ libs/notebook_helpers.py
@@ -7,7 +7,7 @@ from libs.datasets import dataset_utils
 from libs.datasets.sources import covid_tracking_source
 from libs.datasets.sources import nytimes_dataset
 from libs.datasets.sources import texas_hospitalizations
-from libs.datasets.sources import can_scraper_providers
+from libs.datasets.sources import can_scraper_local_dashboard_providers
 
 _logger = logging.getLogger(__name__)
 
@@ -35,7 +35,7 @@ def load_data_sources_by_name() -> Dict[str, pd.DataFrame]:
         texas_hospitalizations.TexasHospitalizations,
         covid_tracking_source.CovidTrackingDataSource,
         nytimes_dataset.NYTimesDataset,
-        can_scraper_providers.CANScraperStateProviders,
+        can_scraper_local_dashboard_providers.CANScraperStateProviders,
     ]
     source_map = {}
     for source_cls in sources:
